requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.3            # via django
cachecontrol==0.12.6      # via firebase-admin
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
cffi==1.14.0
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==5.0.3
cryptography==2.8
dj-static==0.0.6
django-cors-headers==3.2.1
django-extensions==2.2.8
django-mysql==3.3.0
django-oauth-toolkit==1.2.0
django==3.0.3
djangorestframework-jsonapi==3.1.0
djangorestframework==3.11.0
drf-writable-nested==0.5.4
firebase-admin==3.2.1
google-api-core[grpc]==1.16.0  # via firebase-admin, google-cloud-core, google-cloud-firestore
google-api-python-client==1.7.11  # via firebase-admin
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.11.0       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.3.0  # via google-cloud-firestore, google-cloud-storage
google-cloud-firestore==1.6.2  # via firebase-admin
google-cloud-storage==1.25.0  # via firebase-admin
google-resumable-media==0.5.0  # via google-cloud-storage
googleapis-common-protos==1.51.0  # via google-api-core
grpcio==1.27.1            # via google-api-core
gunicorn==20.0.4
httplib2==0.17.0          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
inflection==0.3.1
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
markupsafe==1.1.1         # via jinja2
msgpack==0.6.2            # via cachecontrol
mysqlclient==1.4.6
oauthlib==3.1.0
pip-tools==4.4.1
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.19           # via cffi
python-dateutil==2.8.1
pytz==2019.3              # via django, google-api-core, google-cloud-firestore
pyyaml==5.3
requests==2.22.0
rsa==4.0                  # via google-auth
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.9
six==1.14.0               # via cryptography, django-extensions, firebase-admin, google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, pip-tools, protobuf, python-dateutil
sqlparse==0.3.0           # via django
static3==0.7.0            # via dj-static
stringcase==1.2.0
uritemplate==3.0.1        # via coreapi, google-api-python-client
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
